requirements.txt
Pillow
https://github.com/vishnubob/python-midi/archive/v0.2.3.zip
